Move exception record with pending payments tests to a separate class
